First round of refactoring. Moving code around
